+++ packages/@codotype/ui/src/components/BlueprintEditor/store/schemaModule.js
@@ -1,5 +1,6 @@
 import inflateMeta from '@codotype/util/lib/inflateMeta'
 import sanitizeLabel from '@codotype/util/lib/sanitizeLabel'
+import { inflate } from '@codotype/util/lib/inflate'
 import { buildDefault } from '@codotype/util/lib/buildDefault'
 import { DEFAULT_SCHEMA, SCHEMA_SOURCE_USER } from '@codotype/types/lib/default_schema'
 import formModule from '../../../store/lib/formModule'
@@ -76,6 +77,11 @@ export default {
   getters: {
     selectedModel: state => {
       return state.collection.items.find(i => i.id === state.selectedModel.id)
+    },
+    inflatedSelectedModel: state => {
+      // NOTE - we inflate the entire project here - not efficient, should be cleaned up
+      return inflate({ blueprint: { schemas: state.collection.items }})
+      .schemas.find(s => s.id === state.selectedModel.id)
     },
     defaultObject: state => {
       const schema = state.collection.items.find(i => i.id === state.selectedModel.id)